refactor: split provider and vertexai model files
